requirements.txt
flask
numpy
gunicorn
pickle-mixin
scikit-learn
mysql-connector-python
python-dotenv
Pillow